New RDADynamoHelper class
All dynamo initialization routines have been moved to this class, for now
